requirements.txt
asgiref==3.7.2
Django==4.2.3
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
